fix: id 타입 변경
- fcmTokenId 타입을 int에서 Long으로 변경

+++ src/main/java/com/acc/somsomparty/notification/entity/FcmToken.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class FcmToken {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int fcmTokenId; // id
+    private Long fcmTokenId; // id
 
     private String fcmToken;    // fcm 토큰
